Fixing cpu offload issue

+++ 04-fully-sharded-data-parallel/train_llm.py
@@ -74,10 +74,7 @@ def main():
     fsdp_config = dict(
         reshard_after_forward=True,
         offload_policy=CPUOffloadPolicy() if args.cpu_offload == "on" else None,
-        mp_policy=MixedPrecisionPolicy(
-            param_dtype=torch.bfloat16,
-            reduce_dtype=torch.float32,
-        ),
+        mp_policy=MixedPrecisionPolicy(param_dtype=dtype, reduce_dtype=torch.float32),
     )
     for decoder in model.model.layers:
         fully_shard(decoder, **fsdp_config)
